add custom hook initial state and context
