--- src/components/homepage.js
@@ -1,3 +1,0 @@
-const HomePage = () => {};
-
-export default HomePage;
