+++ src/pages/_app.js
@@ -1,7 +1,16 @@
+import AppContext from '@context/appcontext';
+import useInitialState from '@hooks/useInitialState';
+import Header from '@components/header';
 import '@styles/globals.css';
 
 function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />;
+  const initialState = useInitialState();
+  return (
+    <AppContext.Provider value={initialState}>
+      <Header />
+      <Component {...pageProps} />;
+    </AppContext.Provider>
+  );
 }
 
 export default MyApp;
